Change deployment manager id query parameter to uuid
The `deploymentManagerId` query parameter is expected to aligned
with the similarly named attribute of the object model response
for the API definition.  That is, it should be a UUID value and
an OpenAPI compliant API definition should validate it as such
therefore anything other than a UUID would result in a 400
rather than a 404.

Signed-off-by: Allain Legacy <alegacy@redhat.com>